refactor: 대기 예약 관련 로직 컨트롤러를 별도 클래스로 분리
